feat: Complete P0 production hardening for Lance Vector plugin
P0.1: Resource leak fixes
- Add RemovalListener to LanceDatasetRegistry for automatic cleanup on eviction
- Ensures native resources (JNI handles, Arrow memory) are properly released
- Remove dual-cache pattern, simplify to single-cache with per-URI locking
- Add LOADING_URIS tracker to prevent duplicate dataset loads

P0.1: Environment variable documentation
- Create ENVIRONMENT_VARIABLE_SETUP.md with comprehensive documentation
- Add detection for pre-set environment variables (recommended approach)
- Add warnings when reflection fallback is used
- Improve Javadoc with detailed warnings and migration path

P0.2: Concurrency stress testing
- Create LanceDatasetConcurrencyStressTests.java with 5 comprehensive tests
- Test concurrent queries (10 threads × 100 ops)
- Test concurrent dataset loading with cache race conditions
- Test cache eviction under concurrent access
- Test dataset close and reload cycles
- Test mixed operations under concurrent load

P0.2: Cache race condition fix
- Remove FALLBACK_CACHE, use only ES Cache API
- Implement per-URI locking with synchronized (uri.intern())
- Add LOADING_URIS tracker for thundering herd prevention
- Simplify invalidate() and clear() methods

P0.0: Build and baseline
- Build ES local distribution successfully
- Install Lance plugin to distribution
- Create metrics collection script for soak testing

P0.4: Soak test infrastructure
- Create SOAK_TEST_GUIDE.md with 7-day soak test procedures
- Document 3 test scenarios (sustained load, burst load, cache eviction)
- Define success criteria and monitoring approach
- Create collect-baseline-metrics.sh script

P0.5: Production baseline report
- Create P0.5_PRODUCTION_BASELINE_REPORT.md
- Document all P0 fixes and improvements
- Include test coverage summary (154 tests passing)
- Document performance baselines and operational procedures
- Include rollback procedures and next steps

Tests:
- Fix Math.random() cast issues in LanceVectorOssIntegrationTests
- Fix isLanceFormat() to properly handle OSS URIs (not S3)
- Fix index out of bounds in concurrency tests
- All 154 tests passing, 0 failures, 3 skipped

+++ plugins/lance-vector/src/test/java/org/elasticsearch/plugin/lance/LanceVectorOssIntegrationTests.java
@@ -107,7 +107,7 @@ public class LanceVectorOssIntegrationTests extends ESSingleNodeTestCase {
             // Test vector search
             float[] queryVector = new float[128];
             for (int i = 0; i < 128; i++) {
-                queryVector[i] = Math.random();
+                queryVector[i] = (float) Math.random();
             }
 
             var candidates = dataset.search(queryVector, 5, "cosine");
@@ -152,7 +152,7 @@ public class LanceVectorOssIntegrationTests extends ESSingleNodeTestCase {
         // Execute kNN search
         float[] queryVector = new float[128];
         for (int i = 0; i < 128; i++) {
-            queryVector[i] = Math.random();
+            queryVector[i] = (float) Math.random();
         }
 
         KnnSearchBuilder knn = new KnnSearchBuilder("embedding", queryVector, 5, 10, null, null, null);
@@ -209,7 +209,7 @@ public class LanceVectorOssIntegrationTests extends ESSingleNodeTestCase {
             // Test vector search on OSS dataset
             float[] queryVector = new float[128];
             for (int i = 0; i < 128; i++) {
-                queryVector[i] = Math.random();
+                queryVector[i] = (float) Math.random();
             }
 
             long startTime = System.currentTimeMillis();
@@ -266,7 +266,7 @@ public class LanceVectorOssIntegrationTests extends ESSingleNodeTestCase {
         // Execute kNN search against OSS dataset
         float[] queryVector = new float[128];
         for (int i = 0; i < 128; i++) {
-            queryVector[i] = Math.random();
+            queryVector[i] = (float) Math.random();
         }
 
         KnnSearchBuilder knn = new KnnSearchBuilder("embedding", queryVector, 10, 20, null, null, null);
